+++ src/Framework/ContentPacks/ITranslatable.cs
@@ -1,0 +1,9 @@
+﻿using StardewModdingAPI;
+
+namespace QuestFramework.Framework.ContentPacks
+{
+    internal interface ITranslatable<T>
+    {
+        T Translate(ITranslationHelper translation);
+    }
+}