+++ src/Framework/ContentPacks/Model/Content.cs
@@ -1,5 +1,6 @@
 ﻿using Newtonsoft.Json.Linq;
 using QuestFramework.Offers;
+using QuestFramework.src.Framework.ContentPacks;
 using StardewModdingAPI;
 using System;
 using System.Collections.Generic;
@@ -9,13 +10,23 @@ using System.Threading.Tasks;
 
 namespace QuestFramework.Framework.ContentPacks.Model
 {
-    class Content
+    internal class Content : ITranslatable<Content>
     {
         public ISemanticVersion Format { get; set; }
-
         public List<QuestData> Quests { get; set; }
         public List<QuestOffer<JObject>> Offers { get; set; } = new List<QuestOffer<JObject>>();
 
-        internal IContentPack owner { get; set; }
+        internal IContentPack Owner { get; set; }
+
+        public Content Translate(ITranslationHelper translation)
+        {
+            return new Content()
+            {
+                Format = this.Format,
+                Quests = this.Quests.Select(q => q.Translate(translation)).ToList(),
+                Offers = this.Offers.Select(o => TranslationUtils.TranslateOffer(translation, o)).ToList(),
+                Owner = this.Owner,
+            };
+        }
     }
 }